Made some more Minor adjustments across the Login, SignUp, and Forgot Password page while fixing minor bugs here and there.

+++ public/scripts/controllers/login.js
@@ -17,19 +17,18 @@ app.controller('loginController', function($scope, $http, localStorageService, $
   console.log("loginController Running");
 
   //root Scope Variables
-  $rootScope.LoginTF = 0;
+  $rootScope.LoginTF = 0; //Used within Index to Hide the Header
 
   //Scope Variables
-  $scope.LoginClicked = false;
   $scope.rememberTF = false;
+  $scope.ErrorMessage = "";
+  $scope.BadLogin = false; //Used for Errors on the html Page
   $scope.localLoginauthUser = localStorageService.get('LoginauthUser');
    if($scope.localLoginauthUser != null){
-     //console.log("Fill in Email and Password");
+     console.log("Fill in Email and Password");
      $scope.object = localStorageService.get('LoginauthUser');
      $scope.Email = $scope.object.Email;
      $scope.Password = $scope.object.Password;
-     //console.log("Email: ", $scope.Email);
-     //console.log("Password: ", $scope.Password);
      $scope.rememberTF = true;
    }
 
@@ -37,39 +36,48 @@ app.controller('loginController', function($scope, $http, localStorageService, $
   $scope.login = function(Email, Password){
     console.log("login function called", Email, Password);
     var data = {"Email": Email, "Password": Password};
-    $scope.LoginClicked = true;
-    var message="";
-    //Login Post
-    $http.post("/api/Login/", data)
-    .then(function(data, response) {
-      console.log("data: ", data);
-      if(data.data.authorizedTF)
-      {
-        console.log("Good Login");
-        if($scope.rememberTF == true)
+    if(Email == null || Password == null)
+    {
+      $scope.BadLogin = true; //Used for Errors on the html Page
+      $scope.ErrorMessage = "All Input fields need to be filled out properly!";
+      return;
+    }
+    else {
+      //Login Post
+      $http.post("/api/Login/", data)
+      .then(function(data, response) {
+        console.log("data: ", data);
+        if(data.data.authorizedTF)
         {
-          //console.log("Save Login authUser", data.data);
-          localStorageService.set('LoginauthUser', data.data);
+          console.log("Good Login");
+          if($scope.rememberTF == true)
+          {
+            //console.log("Save Login authUser", data.data);
+            localStorageService.set('LoginauthUser', data.data);
+          }
+          else
+          {
+            //console.log("Clear Login authUser");
+            localStorageService.remove('LoginauthUser');
+          }
+          $scope.IncorrectInputTF = 1;
+          console.log("Logging In...");
+          localStorageService.set('authUser',data.data);
+          $rootScope.authUser = data.data;
+          $scope.BadLogin = false; //Used for Errors on the html Page
+          $rootScope.LoginTF = 1;
+          //console.log("LoginTF: ", $rootScope.LoginTF);
+          $location.path( "/tasks" );
         }
         else
         {
-          //console.log("Clear Login authUser");
-          localStorageService.remove('LoginauthUser');
+          console.log("Bad Login, AuthToken Cannot be Set");
+          $scope.ErrorMessage = "Your Username or password was incorrect. Please Try Again";
+          $scope.BadLogin = true; //Used for Errors on the html Page
+          return;
         }
-        $scope.IncorrectInputTF = 1;
-        console.log("Logging In...");
-        localStorageService.set('authUser',data.data);
-        $rootScope.authUser = data.data;
-        $rootScope.LoginTF = 1;
-        //console.log("LoginTF: ", $rootScope.LoginTF);
-        $location.path( "/tasks" );
-      }
-      else
-      {
-        console.log("Bad Login, AuthToken Cannot be Set");
-        console.log("Your Username or password was incorrect. Please Try Again");
-      }
-    });
+      });
+    }
   };
 
   //SignUp Function
